requirements.txt

# reading environment vars:
python-dotenv

# web application:
Flask #==2.1.1

# auth (login with google via oauth):
Authlib #==1.0.0

# database (google sheets):
gspread
#google-auth==2.16.2
#google-auth-oauthlib==1.0.0


# production web server:
gunicorn

# testing:
pytest
